+++ activejob/CHANGELOG.md
@@ -1,6 +1,7 @@
-*   Skip logging rescuable exceptions which are defined in rescue_from/retry_on/discard_on.
+*   Change the `perform` instrumentation to wrap `perform_now` instead of the `perform` method to stop logging rescuable exceptions defined with `rescue_from`.
 
     *Hu Hailin*
+
 
 
 Please check [6-1-stable](https://github.com/rails/rails/blob/6-1-stable/activejob/CHANGELOG.md) for previous changes.